feat(requirements): 添加FlashIPA相关依赖并更新安装说明
✅ 主要改动：
- 在requirements.txt中新增beartype和jaxtyping依赖
- 更新FlashAttention和FlashIPA的安装说明，提供详细的验证步骤

🔧 其他改进：
- 确保依赖项和安装说明的准确性，以支持FlashIPA的功能

+++ src/stage1/models/ipa.py
@@ -78,7 +78,7 @@ class FlashIPAModuleConfig:
             no_qk_points=self.no_qk_points,
             no_v_points=self.no_v_points,
             use_flash_attn=True,
-            attn_dtype='bf16',
+            attn_dtype='fp16',  # headdim_eff=676 > 256, 必须用fp16
         )
